Final Solution
I tried getting this to say something if you answer yes, but the programs still continues to run, but everything I put it wouldn't make the program stop
it was still showing get a new router even after I said yes to it being fixed.  indentation was incorrect. fixed that.
However when I say yes to step 3 it still gives ask me if that fixed the problem.